Include API call tracing regressions in the existing Python suite
